user authentication

+++ TODOS/router/todos.py
@@ -4,6 +4,7 @@ from fastapi import APIRouter,Depends,HTTPException,status,Path
 from database import SessionLocal
 from models import Todos
 from pydantic import BaseModel,Field
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -15,7 +16,7 @@ def get_db():
         db.close()
         
 db_dependency = Annotated[Session,Depends(get_db)]
-
+user_dependency = Annotated[dict,Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -24,28 +25,44 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(db:db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency,db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def read_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user:user_dependency,
+                    db:db_dependency,
+                    todo_id:int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model:
         return todo_model
     raise HTTPException(status_code=404,detail="To Do not found")
 
 @router.post("/todo",status_code=status.HTTP_201_CREATED)
-async def create_todo(db:db_dependency,insert_todo: TodoRequest):
-    todo_model = Todos(**insert_todo.model_dump())
+async def create_todo(user: user_dependency,
+                      db:db_dependency,
+                      insert_todo: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model = Todos(**insert_todo.model_dump(),
+                       owner_id=user.get('id'))
     
     db.add(todo_model)
     db.commit()
     
 @router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,
-                      todo_request: TodoRequest,
-                      todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todouser(user: user_dependency,
+                          db:db_dependency,
+                          todo_request: TodoRequest,
+                          todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if not todo_model:
         raise HTTPException(status_code=404,detail='To Do Not Found')
     todo_model.title = todo_request.title
@@ -57,9 +74,13 @@ async def update_todo(db:db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,
+async def delete_todo(user: user_dependency,
+                      db:db_dependency,
                       todo_id: int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if not todo_model:
         raise HTTPException(status_code=404,detail="Id Not Found")
     db.delete(todo_model)
